+++ backend/src/app.module.ts
@@ -8,8 +8,8 @@ import { AuthModule } from './auth/auth.module';
 import { BreedModule } from './breed/breed.module';
 import { DatabaseModule } from './common/database/database.module';
 import { HealthController } from './common/health/health.controller';
-import { LoggingInterceptor } from './common/interceptor/logging.interceptor';
-import { PrometheusInterceptor } from './common/interceptor/prometheus.interceptor';
+import { LoggingInterceptor } from './common/interceptors/logging.interceptor';
+import { PrometheusInterceptor } from './common/interceptors/prometheus.interceptor';
 import { WinstonLoggerModule } from './common/logger/winstonLogger.module';
 import { DogWalkDayModule } from './dog-walk-day/dog-walk-day.module';
 import { DogsModule } from './dogs/dogs.module';
